+++ lib/src/telegram/models/abstracts/input_media.dart
@@ -23,5 +23,5 @@ abstract class InputMedia {
   });
 
   /// Converts a [InputMedia] to a [Map] for JSON encoding.
-  Map<String, dynamic> toJson();
+  Map<String, dynamic> getValue([String? field, String? thumb]);
 }