+++ lib/src/telegram/models/inline_query_result_cached_document.dart
@@ -79,11 +79,13 @@ class InlineQueryResultCachedDocument extends InlineQueryResult {
       replyMarkup: json['reply_markup'] == null
           ? null
           : InlineKeyboardMarkup.fromJson(
-              json['reply_markup'] as Map<String, dynamic>),
+              json['reply_markup'] as Map<String, dynamic>,
+            ),
       inputMessageContent: json['input_message_content'] == null
           ? null
           : InputMessageContent.fromJson(
-              json['input_message_content'] as Map<String, dynamic>),
+              json['input_message_content'] as Map<String, dynamic>,
+            ),
     );
   }
 }