+++ lib/src/televerse/models/models.dart
@@ -5,6 +5,7 @@ library televerse.models;
 
 import 'dart:async';
 import 'dart:io' as io;
+import 'dart:typed_data';
 import 'package:televerse/telegram.dart';
 import 'package:televerse/televerse.dart';
 
@@ -18,3 +19,4 @@ part 'mention.dart';
 part 'custom_emoji.dart';
 part 'handler_scope.dart';
 part 'inline_menu_data.dart';
+part 'error.dart';