+++ lib/televerse.dart
@@ -24,7 +24,7 @@ import 'dart:async';
 import 'dart:convert';
 import 'dart:io' as io;
 import 'dart:math';
-import 'package:http/http.dart';
+import 'package:dio/dio.dart';
 import 'package:televerse/telegram.dart';
 import 'package:televerse/televerse.dart';
 import 'package:televerse/src/utils/http.dart';
@@ -49,6 +49,7 @@ part 'src/televerse/filters/filters.dart';
 part 'src/televerse/markups/inline_menu.dart';
 part 'src/televerse/markups/menu.dart';
 part 'src/televerse/markups/keyboard_menu.dart';
+part 'src/televerse/models/multipart_helper.dart';
 
 /// Conversation API
 part 'src/televerse/conversation/conversation.dart';